cast class tensor to Long dtype

+++ bugan/modelsPL.py
@@ -1185,7 +1185,10 @@ class BaseModel(pl.LightningModule):
                 )
                 # combine z and c
                 z = self.merge_latent_and_class_vector(
-                    z, torch.Tensor(c), num_classes, embedding_fn=embedding_fn
+                    z,
+                    torch.Tensor(c).to(torch.int64),
+                    num_classes,
+                    embedding_fn=embedding_fn,
                 )
             else:
                 # generate noise vector
